From 8a0ffc3dfed12b4663dcbe5fccf6c11215830301 Mon Sep 17 00:00:00 2001 From: Conn O'Griofa <connogriofa@gmail.com> Date: Wed, 7 Nov 2012 19:29:04 -0200 Subject: [PATCH] libhardware_legacy/wifi: add workaround for ath6kl driver
The ath6kl driver requires the interface to be brought up in order
for scanning to function. Without this change, the supplicant fails
on first activation of wifi, and will only connect successfully
after multiple retries. This workaround is also necessary to fix
Wi-Fi hotspot mode (in conjunction with another change to
system/netd).

Fixes Tenderloin WiFi

Change-Id: I71fce6017844c02e07b8c02e269d41071cd974ac